Rename `test/test.rs` to `test/utils.rs`
